debug: fixed autograd
so the problem was in the derivative of complex functions.
see https://pytorch.org/docs/stable/notes/autograd.html#complex-autograd-doc for details.
the TMM itself is holomorphic, so what we did was acceptable (inserting \partial M/\partial x inside TMM)
But the mapping from E to, say, R is not holomorphic. Therefore we need Wirtinger calculus
to compose the correct "chain rule"